+++ src/main/java/org/example/model/Customer.java
@@ -8,7 +8,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_id")
-    private Long customerId;
+    private int customerId;
 
     @Column(name = "customer_name")
     private String customerName;
@@ -28,11 +28,11 @@ public class Customer {
 
     //Getters and setters
 
-    public Long getCustomerId() {
+    public int getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(Long customerId) {
+    public void setCustomerId(int customerId) {
         this.customerId = customerId;
     }
